Implement features

+++ app/Http/Controllers/UsersController.php
@@ -7,6 +7,7 @@ use App\Http\Traits\ResponseTrait;
 use App\Http\Requests\ValidateUser;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class UsersController extends Controller
 {
@@ -19,47 +20,74 @@ class UsersController extends Controller
 
     public function getUsers(){
 
-        $users =  $this->user->index();
-
-        if(!$users) return ResponseTrait::sendResponse(null,0,'Failed',422);
-
-        return ResponseTrait::sendResponse($users,1,'Success',200);
+        $authUser = Auth::User();
+        // dd($authUser);
+        if($authUser['role'] == User::ADMIN){
+            $users =  $this->user->index();
+            if(!$users) return ResponseTrait::sendResponse(null,0,'Failed',422);
+            return ResponseTrait::sendResponse($users,1,'Success',200);
+        }else{
+            return ResponseTrait::sendResponse(null,2,'Restricted User',200);
+        }
+        
     }
 
     public function getUserById($id){
 
-        $users =  $this->user->show($id);
+        $authUser = Auth::User();
+        // dd($authUser);
+        if($authUser['role'] == User::ADMIN){
+            $users =  $this->user->show($id);
 
-        if(!$users) return ResponseTrait::sendResponse(null,0,'Failed',422);
+            if(!$users) return ResponseTrait::sendResponse(null,0,'Failed',422);
 
-        return ResponseTrait::sendResponse($users,1,'Success',200);
+            return ResponseTrait::sendResponse($users,1,'Success',200);
+        }else{
+            return ResponseTrait::sendResponse(null,2,'Restricted User',200);
+        }
     }
 
     public function storeUser(ValidateUser $data){
-        $val_data = $data->validated();
-        // if ($val_data->fails()) {
-        //     return ResponseTrait::sendResponse($val_data->errors()->all(),0,'Failed',422);
-        // }
-        $users = $this->user->store($val_data);
 
-        if(!$users) return ResponseTrait::sendResponse(null,0,'Failed',422);
+        $authUser = Auth::User();
+        // dd($authUser);
+        if($authUser['role'] == User::ADMIN){
+            $val_data = $data->validated();
+            $users = $this->user->store($val_data);
 
-        return ResponseTrait::sendResponse($users,1,'Success',200);
+            if(!$users) return ResponseTrait::sendResponse(null,0,'Failed',422);
+
+            return ResponseTrait::sendResponse($users,1,'Success',200);
+        }else{
+            return ResponseTrait::sendResponse(null,2,'Restricted User',200);
+        }
     }
 
     public function updateUser(Request $data, $id){
+        $authUser = Auth::User();
+        // dd($authUser);
+        if($authUser['role'] == User::ADMIN){
 
-        $users =  $this->user->updateUser($data, $id);
+            $users =  $this->user->updateUser($data, $id);
 
-        if(!$users) return ResponseTrait::sendResponse($data,0,'Failed',422);
+            if(!$users) return ResponseTrait::sendResponse($data,0,'Failed',422);
 
-        return ResponseTrait::sendResponse($users,1,'Success',200);
+            return ResponseTrait::sendResponse($users,1,'Success',200);
+        }else{
+            return ResponseTrait::sendResponse(null,2,'Restricted User',200);
+        }
     }
 
     public function deleteUser($id){
-        $users =  $this->user->deleteUser($id);
-        if(!$users) return ResponseTrait::sendResponse($users,0,'Failed',422);
+        $authUser = Auth::User();
+        // dd($authUser);
+        if($authUser['role'] == User::ADMIN){
+            $users =  $this->user->deleteUser($id);
+            if(!$users) return ResponseTrait::sendResponse($users,0,'Failed',422);
 
-        return ResponseTrait::sendResponse($users,1,'Success',200);
+            return ResponseTrait::sendResponse($users,1,'Success',200);
+        }else{
+            return ResponseTrait::sendResponse(null,2,'Restricted User',200);
+        }
     }
 }
